Added test for package imports
Added test to see if gender_novels and dh_testers are installed. This will be run with python setup.py test after the first install of the gender_novels package. I don't fully trust pip to install dh_testers from github and this gives students a simple test to see if the installation succeeded.