Edited find_quoted_text
'Test case: "Miss A.E.--," [...] "a quote"'
now passes
However, I changed the expected result to be from ['"Miss A.E.--"', '"a quote"'] -> ['"Miss A.E.--," "a quote"'] because if it's trying to return cohesive quotes, then it should return broken up ones as full sentences (i.e. 'Test case: "A cat," she said. "is soft." should return ['"A cat," "is soft."'], not ['"A cat,"','"is soft."']
NOTE: would like to speak with whoever added this code more to makes sure I'm not just changing things w/o their consent