Added information to Copyright Info markdown file
Added a Summary section detailing the most critical information to be aware of and a Finer points section with more in-depth points. I have put in sections on US Copyright Law and Project Gutenberg under the Finer points section, and will add HathiTrust, additional information, and edit the text of the document later.