Replaced Joyce's Dubliners with Joyce's Ulysses
Replaced the .txt copy of James Joyce's short story collection "Dubliners" with his novel "Ulysses." "Dubliners" is a short story collection and was therefore deemed problematic, at least for our current scope of study.

The copy of Ulysses was taken from Project Gutenberg and is in the most recent year available for our study.