Added Markdown File for "About Us" Stuff
This is a markdown file that contains the first draft for the information concerning things like "about us" pages and the eggling of a description of our gender novels project. It also has a link to a Google Doc so that people can make suggestions for me to edit/change things if they find errors or simply want to add more info to it.