Removed six books published outside our date range (1776-1922)
Four of the books in our sample corpus were published after 1922 and are therefore outside of our study range (and also possibly not in the public domain). Two of the books in our sample corpus were published before 1776 and deemed too early.

The six books are:
"The Strange Voyage and Adventures of Domingo Gonzales, to the World in the Moon" by Godwin, Francis (1768)
"The Deserted Village" by Goldsmith, Oliver (1770)
"The Merriweather Girls and the Mystery of the Queen's Fan" by Edholm, Lizetter (1932)
"All Jackson's Children" by Galouye, Daniel F. (1957)
"Damned If You Don't" by Garrett, Randall (1960)
"Subject to Change" by Goulart, Ron (1960)