Basic pronoun frequency analysis code additions
I added doctests to all the methods except for the first two. The first two require the ability to make a subcorpus to do a doctest unless the doctest is going to take a couple minutes.

Also, a new analysis has been added to compare subject pronouns vs object pronoun use for masculine and feminie pronouns.

The statements to actually run all of this code and print the results are at the bottom. They are commented out so you can choose what exactly you would like to run and then uncomment that bit.