Removed _remove_table_of_contents()
After a long discussion with Ife, Dina, and Stephan, it was concluded that removing the table of contents is not worth it based on the various different formats and edge cases. Rather than dealing with ToCs on the data side of things, it is far easier to instead do a try/catch on the analysis side of things, where if the sentence is unparsable and occurs early in the novel (and repeatedly occurs in a row, as ToCs are) the line is thrown out.