Setup multi-page structure with Jinja (#183)
* Moved navbar horizontally

Moved navbar horizontally to add dropdown functionality, but that dropdown is still a work in progress. Also, commented out the pandas functions I was working on for CSV conversion because it was changing a local file on every execution of app.py

* Formatted horizontal navbar a bit more

Changed a few things in the norizontal navbar like where the content is justified and what exact tabs are in it

* Setup HTML and CSS for multiple page structure

Setup the HTML and CSS files already written to be in a multi-page system instead of a single page with dynamic content rendering.

* Converted all content to the multi-page structure

Converted all HTML content written so far into the multi-page structure with a horizontal navigation bar that is now fully functional due to template rendering in app.py. Additionally, added a test page tab so other lab members can test out the markdown conversion script on their own markdowns and see the result on the site