Reverted indicated lower bound of year range to 1770
Our original year bound was 1770-1922; midway through the project we shifted bounds to 1776-1922, but it seems this shift was not completely updated. The corpus and other analyses were generated with the original bound of 1770, and it is impractical to regenerate the corpus or delete the removed years, so we are reverting to the original bound of 1770.