Grammar Analysis Write-Up in Markdown Form
This is the writeup for the grammar analysis that has been converted into a markdown file for the website. It will have some parts added to it within the next few days (i.e. added examples at the end for differences between male and female pronoun adjective pairings), but is a start for pushing it to the website.

Additionally, there are minor formating changes to the overview page in order to make the landing page a little cleaner.